backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.30
pydantic==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.3.0
feedparser==6.0.11
beautifulsoup4==4.12.3
httpx==0.27.0
google-generativeai==0.7.0
python-dotenv==1.0.1
